Pocket Health
OBJECTIVE:
Health monitoring app development is the best way to improve your health in 
your comfort zone. The objective of this project is to develop a mobile platform 
that collect data from the individual and detects the possible health issues using 
symptoms.

INTRODUCTION:
Technology and smartphones may cause harm but it totally depends on how you 
utilize it. Health monitoring app development is the best way to improve your 
health without spending thousand bucks.
With the health monitoring app development there are a lot of features that 
make the application more effective and easy to use. Health app works as a 
reminder and in the emergence we can make the most benefit out of it.

PURPOSE:
As we all agree that Health Is Wealth if you have a healthy body and a healthy 
mind then only you live a good life. The purpose of a health monitoring app is 
not limited to your daily cardio routine but it solves many more purposes such 
as SOS message to emergency contact when vigorous movement is detected, 
reminder setting for regular medicine intake schedule, health chat bot with fully 
functional API and many more. 
According to the current lockdown situation all over the world where we have 
to maintain social distancing due to the COVID-19 (coronavirus) health 
monitoring app is a boon. We are so fortunate that in this quarantine and in the 
complete isolating situation, with the help of technology you can pay attention 
to our health with just one health app.

TECHNOLOGY REQUIREMENTS:
1. JAVA
2. XML
3. SQLite for database

KEY FEATURES:
1. SOS- Whenever vigorous shaking is detected by the app, it will send an 
emergency message to your already selected emergency contacts. If you 
are in an accident or you had fainted of tiredness, the shaking of your 
phone is detected and then the SMS is sent.
2. Notification and reminders- We all agree that at some point we always 
ignore our health because our own tardiness we always need someone to 
push us, health monitoring apps having notification and reminders will 
help us to push and carry our daily health routine.
3. Chatbot- While health monitoring app development we make sure we 
develop chatbot features, it is the easiest and essential feature in the 
health monitoring app if you have any doubt you can just ping your query 
and you will get your answer.
4. Medicine reminder- It is very crucial for any patient to take their medicine 
on time health monitoring app can help the person to take their medicine 
on time by setting reminders of what medicine to take at what time.